Organize ML database types and fixtures by feature
