+++ rust/crates/photos/src/ml_db/constants.rs
@@ -1,9 +1,5 @@
 pub const FACE_ML_VERSION: i64 = 1;
-pub const CLIP_ML_VERSION: i64 = 1;
 pub const PET_ML_VERSION: i64 = 1;
-
-pub const CLIP_EMBEDDING_DIMENSIONS: usize = 512;
-pub const CLIP_EMBEDDING_BYTES_LENGTH: i64 = CLIP_EMBEDDING_DIMENSIONS as i64 * 4;
 
 pub const LAPLACIAN_HARD_THRESHOLD: f64 = 10.0;
 pub const LAPLACIAN_SOFT_THRESHOLD: f64 = 50.0;
