--- rust/crates/photos/src/ml_db/error.rs
@@ -1,13 +1,0 @@
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    #[error(transparent)]
-    Database(#[from] crate::db::Error),
-    #[error("{0}")]
-    Codec(String),
-    #[error(transparent)]
-    Io(#[from] std::io::Error),
-    #[error("{0}")]
-    InvalidArgument(String),
-}
-
-pub type Result<T> = std::result::Result<T, Error>;
