+++ rust/crates/photos/src/ml_db/filedata.rs
@@ -2,9 +2,24 @@ use std::collections::{HashMap, HashSet};
 
 use crate::db::optional_parameter;
 
-use super::MlDb;
-use super::error::Result;
-use super::types::{FdStatus, PreviewInfo};
+use super::{MlDb, Result};
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct FdStatus {
+    pub file_id: i64,
+    pub user_id: i64,
+    pub data_type: String,
+    pub size: i64,
+    pub object_id: Option<String>,
+    pub object_nonce: Option<String>,
+    pub updated_at: i64,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct PreviewInfo {
+    pub object_id: String,
+    pub object_size: i64,
+}
 
 impl MlDb {
     pub fn put_fd_status(&self, fd_status_list: &[FdStatus]) -> Result<()> {
@@ -57,25 +72,73 @@ impl MlDb {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
+pub(super) mod tests {
+    use std::collections::{HashMap, HashSet};
 
-    use crate::ml_db::tests::{cases, check_seeded, ids, preview, seeded, status};
+    use super::{FdStatus, MlDb, PreviewInfo};
+    use crate::ml_db::tests::{cases, check, open};
+    use tempfile::TempDir;
+
+    fn status(file_id: i64, data_type: &str, object_id: Option<&str>) -> FdStatus {
+        FdStatus {
+            file_id,
+            user_id: 42,
+            data_type: data_type.to_string(),
+            size: file_id * 10,
+            object_id: object_id.map(str::to_string),
+            object_nonce: None,
+            updated_at: 1000,
+        }
+    }
+
+    fn preview(object_id: &str, object_size: i64) -> PreviewInfo {
+        PreviewInfo {
+            object_id: object_id.to_string(),
+            object_size,
+        }
+    }
+
+    fn ids<const N: usize>(values: [i64; N]) -> HashSet<i64> {
+        HashSet::from(values)
+    }
+
+    pub(in crate::ml_db) fn seed(db: &MlDb) {
+        db.put_fd_status(&[
+            status(1, "vid_preview", Some("obj1")),
+            status(2, "mldata", None),
+            status(3, "vid_preview", Some("obj3")),
+        ])
+        .unwrap();
+    }
+
+    fn seeded() -> (TempDir, MlDb) {
+        let (directory, db) = open();
+        seed(&db);
+        (directory, db)
+    }
 
     #[test]
     fn seeded_file_id_sets() {
-        check_seeded(&cases![
-            "fd data": ids([1, 2, 3]) => |db| db.get_file_ids_with_fd_data(None),
-            "fd mldata": ids([2]) => |db| db.get_file_ids_with_fd_data(Some("mldata")),
-        ]);
+        let (_directory, db) = seeded();
+        check(
+            &db,
+            &cases![
+                "fd data": ids([1, 2, 3]) => |db| db.get_file_ids_with_fd_data(None),
+                "fd mldata": ids([2]) => |db| db.get_file_ids_with_fd_data(Some("mldata")),
+            ],
+        );
     }
 
     #[test]
     fn seeded_video_previews() {
-        check_seeded(&cases![
-            "video previews": HashMap::from([(1, preview("obj1", 10)), (3, preview("obj3", 30))]) =>
-                |db| db.get_file_ids_vid_preview(),
-        ]);
+        let (_directory, db) = seeded();
+        check(
+            &db,
+            &cases![
+                "video previews": HashMap::from([(1, preview("obj1", 10)), (3, preview("obj3", 30))]) =>
+                    |db| db.get_file_ids_vid_preview(),
+            ],
+        );
     }
 
     #[test]
